+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - New function `is_subscribed` to check whether a given user is subscribed to a given list
   - New function `add_member` to add a user to a list
   - New function `create` to create a new list
+  - New function `delete` to delete a list
 - New enum `service::ListMethod` and `list` field in `service::RateLimitStatus` to contain the rate
   limit info for these methods